requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
networkx>=2.6.0
pygraphviz>=1.7.0,<2.0.0

# Data handling and configuration
pyyaml>=5.4.0,<6.0.0
typing-extensions>=4.0.0,<5.0.0

# Progress tracking
tqdm>=4.62.0,<5.0.0

# Machine learning (if needed)
scikit-learn>=0.24.0,<1.0.0

# Development dependencies
pytest>=6.2.0,<7.0.0
jupyter>=1.0.0,<2.0.0
black>=22.0.0,<23.0.0  # Code formatting
flake8>=4.0.0,<5.0.0  # Code linting
mypy>=0.900,<1.0.0    # Type checking